AFB Module added

+++ CIStudies/interface/AnalysisModule.hh
@@ -21,7 +21,7 @@ protected:
   const TObject* getObject(const std::string& name) const {return objects.at(name);}
 
   void makeHistogram(const std::string& name, const std::string& title, int nbins,
-		     int min, int max);
+		     double min, double max);
   TH1* getHistogram(const std::string& name) {return dynamic_cast<TH1*>(getObject(name));}
   const TH1* getHistogram(const std::string& name) const
   {return dynamic_cast<const TH1*>(getObject(name));}
